requirements.txt
torch==1.11.0
numpy==1.19.2
pilgram==1.2.0
piqa==1.1.3
py==1.10.0
matplotlib==3.3.4
opencv-python==4.4.0.46
opencv-python-headless==4.5.3.56
Pillow==8.0.0
